requirements.txt
absl-py==0.2.2
astor==0.6.2
astroid==1.5.3
bleach==1.5.0
cycler==0.10.0
Django==2.0.7
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
enum34==1.1.6
gast==0.2.0
grpcio==1.13.0
html5lib==0.9999999
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
matplotlib==2.1.0
mccabe==0.6.1
nltk==3.3
numpy==1.13.3
pandas==0.23.1
protobuf==3.5.0.post1
pylint==1.7.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2017.3
scipy==1.1.0
seaborn==0.8.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-hub==0.1.0
tensorflow-tensorboard==0.4.0rc3
termcolor==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
